fix(keymap): correct per-key config writes on the RY5088
Two firmware-validated (Ghidra v407) per-key write bugs, both of which silently
misconfigured keys:

1. Per-key trigger config went through SET_KEY_MAGNETISM_MODE (0x1D) — a YC500-only
   command that CommonKBRY5088.js stubs to a no-op — so per-key actuation/release/
   mode edits never landed. Route get/set_key_trigger through the multi-magnetism
   table (subcmd 0x00/0x01/0x07, u16 @ device precision) exactly as the vendor web
   app does. KeyTriggerSettings.actuation/deactuation are now u16; CLI and TUI drop
   the bogus u8/0.1mm scale for the real precision factor.

2. A [0,0,0,0] keymatrix entry emits keycode 0 = silence, and there is NO factory
   fallback for the base layer (the "enabled" byte only marks flash-dirty, it does
   not gate output). So reset-key on a base key and DKS's empty binding-0 were
   silencing keys (this disabled Esc during testing). Base-layer reset now writes
   the position's factory-default keycode; DKS binding-0 preserves the key's current
   layer-0 output; DKS commits once (final binding) + settles, avoiding a flash-save
   readback desync.

Verified on-device (M1 V5 TMR): per-key actuation edits land and read back; reset-
key and set_dks_config never silence a key; dks-roundtrip shows no cross-key
disturbance across repeats.

+++ iot_driver_linux/monsgeek-keyboard/src/magnetism.rs
@@ -128,10 +128,10 @@ impl TriggerSettings {
 pub struct KeyTriggerSettings {
     /// Key matrix index
     pub key_index: u8,
-    /// Actuation point (raw units)
-    pub actuation: u8,
-    /// Deactuation point (raw units)
-    pub deactuation: u8,
+    /// Actuation point (raw u16 firmware units, precision-dependent)
+    pub actuation: u16,
+    /// Deactuation / release point (raw u16 firmware units)
+    pub deactuation: u16,
     /// Base key mode
     pub mode: KeyMode,
     /// Rapid-Trigger flag (orthogonal to `mode`)
